optimize table rendering
