+++ src/components/Spreadsheet/components/Table.tsx
@@ -1,11 +1,21 @@
+import { useState } from 'react'
+import { useTableEvents } from '@/components/Spreadsheet/logic/useTableEvents'
+import { useMountTransition } from '@/logic/useMountTransition'
+import { useMenu } from '@/components/ContextualMenu/logic/useContextMenu'
+import { ContextualMenu } from '@/components/ContextualMenu'
 import { inputTag } from '@/components/Spreadsheet/data/constants'
 
 type PropTypes = {
   readonly children: React.ReactNode
-  readonly onClick: (event: React.MouseEvent | React.KeyboardEvent) => void
 }
 
-export function Table({ children, onClick }: PropTypes) {
+export function TableWithMenu({ children }: PropTypes) {
+  const { selectedElements } = useTableEvents()
+
+  const [openMenu, setOpenMenu] = useState(false)
+  const isMounted = useMountTransition(openMenu)
+  const { setPosition, coords, origin } = useMenu(() => setOpenMenu(true))
+
   const handleContextMenu = (event: React.MouseEvent) => {
     const target = event.target as HTMLElement
     if (target.tagName === inputTag) return
@@ -18,10 +28,19 @@ export function Table({ children, onClick }: PropTypes) {
       className='table-wrapper'
       // listeners to determine contextual menu behavior
       onContextMenu={handleContextMenu}
-      onMouseUp={onClick}
-      onKeyDown={onClick}
+      onMouseUp={setPosition}
+      onKeyDown={setPosition}
     >
       <table aria-label='spreadsheet'>{children}</table>
+      {(openMenu || isMounted) && (
+        <ContextualMenu
+          className={`${isMounted && 'in'} ${openMenu && 'visible'}`}
+          onClose={() => setOpenMenu(false)}
+          coords={coords}
+          origin={origin}
+          selectedItems={selectedElements}
+        />
+      )}
     </section>
   )
 }
